requirements.txt

python-telegram-bot
openai
requests
beautifulsoup4
python-dotenv
